fix: update pagination and table
