--- components/tables/TableInfo.jsx
@@ -1,75 +1,0 @@
-import React from 'react';
-import { BasicTable } from './BasicTable';
-
-const data = [
-  { 'number': '1',
-    'ID': 1234,
-    'kpgz': '02.03.03.11 ОБУСТРОЙСТВ ПОКРЫТИЙ И ЭЛЕМЕНТОВ СОПРЯЖЕНИЯ ТЕРРИТОРИЙ',
-    'spgz': 'Устройство из резиновой в рамках благоустройств территории',
-    'unit': 'Квадраты',
-    'capacity': '318. 00',
-    'price': '3 758, 89',
-    'summa': '1 195 234, 00',
-    'address': '"Пешеходна "Аллея вязо (ЗелАО)',
-  },
-  { 'number': '1',
-    'ID': 1234,
-    'kpgz': '02.03.03.11 ОБУСТРОЙСТВ ПОКРЫТИЙ И ЭЛЕМЕНТОВ СОПРЯЖЕНИЯ ТЕРРИТОРИЙ',
-    'spgz': 'Устройство из резиновой в рамках благоустройств территории',
-    'unit': 'Квадраты',
-    'capacity': '318. 00',
-    'price': '3 758, 89',
-    'summa': '1 195 234, 00',
-    'address': '"Пешеходна "Аллея вязо (ЗелАО)',
-  },
-  { 'number': '1',
-    'ID': 1234,
-    'kpgz': '02.03.03.11 ОБУСТРОЙСТВ ПОКРЫТИЙ И ЭЛЕМЕНТОВ СОПРЯЖЕНИЯ ТЕРРИТОРИЙ',
-    'spgz': 'Устройство из резиновой в рамках благоустройств территории',
-    'unit': 'Квадраты',
-    'capacity': '318. 00',
-    'price': '3 758, 89',
-    'summa': '1 195 234, 00',
-    'address': '"Пешеходна "Аллея вязо (ЗелАО)',
-  },
-  { 'number': '1',
-    'ID': 1234,
-    'kpgz': '02.03.03.11 ОБУСТРОЙСТВ ПОКРЫТИЙ И ЭЛЕМЕНТОВ СОПРЯЖЕНИЯ ТЕРРИТОРИЙ',
-    'spgz': 'Устройство из резиновой в рамках благоустройств территории',
-    'unit': 'Квадраты',
-    'capacity': '318. 00',
-    'price': '3 758, 89',
-    'summa': '1 195 234, 00',
-    'address': '"Пешеходна "Аллея вязо (ЗелАО)',
-  },
-  { 'number': '1',
-    'ID': 1234,
-    'kpgz': '02.03.03.11 ОБУСТРОЙСТВ ПОКРЫТИЙ И ЭЛЕМЕНТОВ СОПРЯЖЕНИЯ ТЕРРИТОРИЙ',
-    'spgz': 'Устройство из резиновой в рамках благоустройств территории',
-    'unit': 'Квадраты',
-    'capacity': '318. 00',
-    'price': '3 758, 89',
-    'summa': '1 195 234, 00',
-    'address': '"Пешеходна "Аллея вязо (ЗелАО)',
-  },
-];
-
-const columns = [
-  { id: 'number', label: '№' },
-  { id: 'ID', label: 'ID' },
-  { id: 'kpgz', label: 'КПГЗ' },
-  { id: 'spgz', label: 'СПГЗ' },
-  { id: 'unit', label: 'Единица измерения' },
-  { id: 'capacity', label: 'Объем' },
-  { id: 'price', label: 'Цена за единицу, руб' },
-  { id: 'summa', label: 'Сумма' },
-  { id: 'address', label: 'Адрес' },
-];
-
-export const TableInfo = () => {
-  return(
-    <BasicTable columns={columns} data={data}/>
-  );
-};
-
-
